exercise store refactor
