+++ components/basecomponents/Exercise/components/ExerciseAudioIdentificationComponent.tsx
@@ -1,4 +1,4 @@
-import { Choice, ExerciseStatus, resolveMethods, resultMethods, useExerciseStore } from '../exerciseStore';
+import { Choice, ExerciseStatus, useExerciseStore } from '../exerciseStore';
 import React, { useRef, useEffect, useImperativeHandle, forwardRef } from 'react';
 import { animation } from '../utils/animation';
 import SoundWaveIcon from 'public/icons/sound-wave.svg';
@@ -10,6 +10,8 @@ import { ChoiceListComponent } from './ChoiceListComponent';
 import { ExerciseContainer } from './ExerciseContainer';
 import { Hint } from './Hint';
 import { MainText } from './MainText';
+import { resolveMethods } from '../utils/resolveMethods';
+import { resultMethods } from '../utils/resultMethods';
 
 interface ExerciseAudioIdentificationComponentProps {
   choices: Choice[];
