+++ components/basecomponents/Exercise/utils/delay.ts
@@ -1,0 +1,1 @@
+export const delay = (ms: number) => new Promise<void>((resolve) => setTimeout(resolve, ms));
